+++ src/test/java/org/fundacionjala/coding/ariel/cesarvigenereencryption/CesarEncryptionTest.java
@@ -1,19 +1,12 @@
-package org.fundacionjala.coding.ariel;
+package org.fundacionjala.coding.ariel.cesarvigenereencryption;
 
-import org.fundacionjala.coding.ariel.cesarvigenereencryption.CesarEncryption;
-import org.fundacionjala.coding.ariel.cesarvigenereencryption.VigenereEncryption;
 import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-/**
- * CesarCrypto class tests.
- */
-public class CesarVigenereEncryptionTest {
+public class CesarEncryptionTest {
     private CesarEncryption cesarEncryption;
-    private VigenereEncryption vigenereEncryption;
-
 
     /**
      * Init the class cesarvigenereencryption.
@@ -21,7 +14,6 @@ public class CesarVigenereEncryptionTest {
     @Before
     public void init() {
         cesarEncryption = new CesarEncryption();
-        vigenereEncryption = new VigenereEncryption();
 
     }
 
@@ -77,44 +69,4 @@ public class CesarVigenereEncryptionTest {
         assertEquals("KAT?A", cesarEncryption.decodeMessage("RHA?H", "-7"));
         assertEquals("P@ATA TA", cesarEncryption.decodeMessage("E@PIP IP", "-15"));
     }
-
-    /**
-     * Test that check the Vigenere encode with special.
-     */
-    @Test
-    public void encodeVigenereMessage() {
-        assertEquals("TXYP XMUNHCRC", vigenereEncryption.encodeMessage("HOLA JALASOFT", "LIMON"));
-        assertEquals("UTXP AHSPD", vigenereEncryption.encodeMessage("HOLA MUNDO", "MELON"));
-        assertEquals("NTFU NT WWRDE ACVROYWY", vigenereEncryption.encodeMessage("TEST TO CHECK VIGENERE", "TOMATE"));
-
-    }
-
-    /**
-     * Test that check the Vigenere decode with special.
-     */
-    @Test
-    public void decodeVigenereMessage() {
-        assertEquals("HOLA JALASOFT", vigenereEncryption.decodeMessage("TXYP XMUNHCRC", "LIMON"));
-        assertEquals("HOLA MUNDO", vigenereEncryption.decodeMessage("UTXP AHSPD", "MELON"));
-        assertEquals("TEST TO CHECK VIGENERE", vigenereEncryption.decodeMessage("NTFU NT WWRDE ACVROYWY", "TOMATE"));
-    }
-
-    /**
-     * Test that check a null and empty input.
-     */
-    @Test
-    public void emptyTextEncodeVigenere() {
-        assertEquals("", vigenereEncryption.encodeMessage("", "HOLA"));
-        assertEquals("", vigenereEncryption.encodeMessage(null, "MUNDO"));
-    }
-
-    /**
-     * Test that check a null and empty input.
-     */
-    @Test
-    public void emptyTextDecodeVIgenere() {
-        assertEquals("", vigenereEncryption.decodeMessage("", "UTXP"));
-        assertEquals("", vigenereEncryption.decodeMessage(null, "WWRDE"));
-    }
-
 }
